made brand selection into component and moved onChange function
